# [Speed]Refine elementwise_mul_op